requirements.txt
click==8.0.3
jsonschema==3.2.0
pydantic==1.10.4
pymongo==4.3.3
python-arango==7.5.5
python-dotenv==0.21.0
sanic==22.12.0
Sanic-Cors==2.2.0
sanic-ext==22.12.0
sanic-routing==22.8.0
sanic-testing==23.3.0
termcolor==2.2.0
redis==4.4.2
requests==2.28.1
urllib3==1.26.13
pycoingecko==3.1.0
PyJWT==2.6.0
websockets==10.4
multithread-parallel-processing==1.0.0
numpy==1.24.2
SQLAlchemy~=1.4.46
psycopg2-binary
pytest==7.2.2
prettytable==3.6.0